requirements.txt
anthropic>=0.25.0
openai>=1.30.0
langchain>=0.2.0
langchain-anthropic>=0.1.15
langchain-openai>=0.1.8
langchain-community>=0.2.0
faiss-cpu>=1.8.0
chromadb>=0.5.0
streamlit>=1.35.0
pandas>=2.2.0
numpy>=1.26.0
faker>=25.0.0
python-dotenv>=1.0.0
plotly>=5.22.0
tiktoken>=0.7.0
sentence-transformers>=3.0.0
